video player into component
