+++ packages/frontend/src/components/Home/index.js
@@ -1,24 +1,8 @@
 /* eslint-disable jsx-a11y/media-has-caption */
-import React, { useEffect, useRef } from 'react';
+import React from 'react';
+import VideoPlayer from '../VideoPlayer';
 
 function Home() {
-  const videoRef = useRef(null);
-  const progressRef = useRef(null);
-
-  useEffect(() => {
-    const video = videoRef.current;
-    const progress = progressRef.current;
-
-    const updateProgress = () => {
-      const value = Math.round((video.currentTime / video.duration) * 100);
-      progress.value = value;
-    };
-
-    video.addEventListener('timeupdate', updateProgress);
-
-    return () => video.removeEventListener('timeupdate', updateProgress);
-  }, []);
-
   return (
     <div className="wrap">
       <div className="header">
@@ -37,16 +21,7 @@ function Home() {
         </div>
 
         <div className="right-col">
-
-          <figure>
-            <video autoPlay muted loop id="video" ref={videoRef}>
-              <source src="/map-assets/one-day-loop.mp4" type="video/mp4" />
-            </video>
-            <figcaption>
-              <progress id="progress" max="100" value="0" ref={progressRef}>Progress</progress>
-            </figcaption>
-          </figure>
-
+          <VideoPlayer currentSelection="/map-assets/one-day-loop.mp4" />
         </div>
       </div>
     </div>
